Fix "no tty sudo" error on centos-5
Prior to this change, if one attempted to run sudo from a shell without
a tty, the redhat-version of sudo would complain with:
    sudo: sorry, you must have a tty to run sudo
This would occur if one was running sudo from a ssh-remote-command or
automated docker instance in a build pipeline.

It is discussed in
https://bugzilla.redhat.com/show_bug.cgi?id=1020147#c9 and the sudoers
optioon that enabled it was decided to be wrong and should be disabled.
Unfortunantely, that fix was never backported to centos5 as it was
likely EOL during this decision.

This change disables the 'requiretty' option in the sudoers file.